🎨  rename assistant_model.dart and update id type to int; add assistant_list.dart with sample assistants
